// src/core/interfaces/keyboard.interface.ts
import { KeyboardSize } from "../types/keyboard-size.type";

export interface IKeyboard {
    piano: {
        whiteKeyNotes: number[];
        blackKeyNotes: number[];
    };
    large: {
        whiteKeyNotes: number[];
        blackKeyNotes: number[];
    };
    medium: {
        whiteKeyNotes: number[];
        blackKeyNotes: number[];
    };
    small: {
        whiteKeyNotes: number[];
        blackKeyNotes: number[];
    };
}